+++ tests/surfaces/test_revolutionsurface3d.py
@@ -103,6 +103,14 @@ class TestRevolutionSurface3D(unittest.TestCase):
         self.assertAlmostEqual(arc.radius, 0.022500000035448893)
         self.assertAlmostEqual(arc.angle, 0.15581712793343738)
 
+    def test_contour3d_to_2d(self):
+        surface = surfaces.RevolutionSurface3D.load_from_file(
+            "surfaces/objects_revolution_tests/revolutionface_surface.json")
+        contour = vmw.Contour3D.load_from_file("surfaces/objects_revolution_tests/revolutionface_contour.json")
+        contour2d = surface.contour3d_to_2d(contour)
+        self.assertTrue(contour2d.is_ordered())
+        self.assertAlmostEqual(contour2d.area(), 0.031887165433924704 * 0.5 * math.pi, 2)
+
 
 if __name__ == '__main__':
     unittest.main()